requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
djangorestframework
Pillow==10.0.1
django-crispy-forms>=1.9.0
crispy_bootstrap5
requests
pyyaml
django-cors-headers